% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_rprofile.R
\name{add_fcuk_to_rprofile}
\alias{add_fcuk_to_rprofile}
\title{Add library(fcuk) to the .Rprofile file}
\usage{
add_fcuk_to_rprofile()
}
\description{
After calling this function, fcuk will be launched everytime you launch
your current R project.
}
\examples{
fcuk::add_fcuk_to_rprofile()
}
