% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{catch_error}
\alias{catch_error}
\title{capture and parse the last error}
\usage{
catch_error(sentence = geterrmessage())
}
\arguments{
\item{sentense}{}
}
\description{
capture and parse the last error
}
