% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun.R
\docType{data}
\name{.}
\alias{.}
\title{Placeholder object for rerun operator}
\format{
An object of class \code{fcuk} of length 1.
}
\usage{
.
}
\description{
Use `.+1` or `.+2` to rerun the last command with a suggested correction.
}
\examples{
\dontrun{
view(iris) # error
. + 1 # returns View(iris)
}
}
\keyword{datasets}
