% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{erreur_correction_propostion}
\alias{erreur_correction_propostion}
\title{Find closest R functions names}
\usage{
erreur_correction_propostion(asked_objet, method = "jaccard", n = 2)
}
\arguments{
\item{asked_objet}{the R object name producing an error}

\item{method}{Method for distance calculation. The default is "jaccard", see \link[stringdist]{stringdist-metrics}.}

\item{n}{number of names corrections to suggest}
}
\description{
Find closest R functions names
}
\examples{
erreur_correction_propostion("iri")
}
