% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{error_analysis}
\alias{error_analysis}
\title{Error Analysis}
\usage{
error_analysis(asked_objet = catch_error(), n = 2)
}
\arguments{
\item{asked_objet}{the R object name producing an error}

\item{n}{number of corrections to propose}
}
\description{
Error Analysis
}
\examples{
fcuk::error_analysis() #last error is analysed
fcuk::error_analysis("view")
fcuk::error_analysis("iri")
}
