% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{error_correction_propostion}
\alias{error_correction_propostion}
\title{Find closest object names}
\usage{
error_correction_propostion(asked_objet, method = "jaccard", n = 2)
}
\arguments{
\item{asked_objet}{the R name producing an error}

\item{method}{Method for distance calculation. The default is "jaccard", see \link[stringdist]{stringdist-metrics}.}

\item{n}{number of corrections to suggest.}
}
\value{
a character vector with the closest neighbors
}
\description{
Analyse a typo and suggests the two closest names.
}
\examples{
error_correction_propostion("iri")
}
