% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{get_all_objects_from_r}
\alias{get_all_objects_from_r}
\alias{get_all_objets_from_r}
\title{Extract the name of all objects loaded in the R environments}
\usage{
get_all_objects_from_r(use_cache = TRUE)

get_all_objets_from_r(use_cache = TRUE)
}
\arguments{
\item{use_cache}{Logical. Use cached results if available (default: TRUE).
Cache is invalidated when environment changes.}
}
\value{
A character vector with the names of all objects contained in the environments
}
\description{
Fetch the name of all objects loaded in the environments (functions, values, data...)
}
\examples{
get_all_objects_from_r()
}
