% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{get_all_objets_from_r}
\alias{get_all_objets_from_r}
\title{Extract all loaded objects names from R environments}
\usage{
get_all_objets_from_r(...)
}
\arguments{
\item{...}{not used}
}
\description{
fetch all R loaded objects names (functions, values, data...)
}
\examples{
get_all_objets_from_r()
}
