% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun.R
\name{get_last}
\alias{get_last}
\title{return corrected instruction}
\usage{
get_last(n = 1)
}
\arguments{
\item{n}{number of correction}
}
\description{
return corrected instruction
}
\examples{
get_last()
}
