% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{init_error_tracker}
\alias{init_error_tracker}
\title{init error tracker}
\usage{
init_error_tracker()
}
\description{
every errors will be analysed
}
\examples{
getOption("error")
fcuk::init_error_tracker()
getOption("error")
}
