% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun.R
\name{+.fcuk}
\alias{+.fcuk}
\title{Rerun with correcyion}
\usage{
\method{+}{fcuk}(x, y)
}
\arguments{
\item{x}{blank object of class fcuk}

\item{y}{an integer corresponding to the position of the proposal to be reused}
}
\description{
Rerun with correcyion
}
\examples{
\dontrun{
view(iris)# error
.+1 # return View(iris)

}


}
