% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{remove_error_tracker}
\alias{remove_error_tracker}
\title{Remove error tracker}
\usage{
remove_error_tracker()
}
\description{
After lauching this function, the errors will no longer be analysed.
}
\examples{
getOption("error")
fcuk::remove_error_tracker()
getOption("error")
}
