% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_rprofile.R
\name{add_fcuk_to_rprofile}
\alias{add_fcuk_to_rprofile}
\title{Add library(fcuk) to the .Rprofile file}
\usage{
add_fcuk_to_rprofile(global = TRUE)
}
\arguments{
\item{global}{Whether to add it to the global .Rprofile (`TRUE`, the default)
or just to the .Rprofile file of your current project (`FALSE`).}
}
\value{
Invisibly returns `NULL`.
}
\description{
After calling this function, fcuk will be launched everytime you launch your
current R project (or any project if you use option `global = TRUE`).
}
\examples{
\dontrun{
fcuk::add_fcuk_to_rprofile()
}
}
