% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_rprofile.R
\name{add_fcuk_to_rprofile}
\alias{add_fcuk_to_rprofile}
\title{add library(fcuk) in the .Rprofile file}
\usage{
add_fcuk_to_rprofile()
}
\description{
add library(fcuk) in the .Rprofile file
}
\examples{
fcuk::add_fcuk_to_rprofile()
}
