% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{catch_error}
\alias{catch_error}
\title{Capture and parse an error message.}
\usage{
catch_error(sentence = geterrmessage())
}
\arguments{
\item{sentence}{an error message to parse}
}
\description{
Capture and parse an error message.
}
\examples{
catch_error()
catch_error("Error: object 'iri' not found\\n")
catch_error("Error: object 'view' not found\\n")


}
