% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{erreur_correction_propostion}
\alias{erreur_correction_propostion}
\title{Find closest R functions}
\usage{
erreur_correction_propostion(car, method = "jaccard")
}
\arguments{
\item{car}{error to check}

\item{method}{Method for distance calculation. The default is "jaccard", see \link[stringdist]{stringdist-metrics}.}
}
\description{
Find closest R functions
}
