% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{error_correction_propostion}
\alias{error_correction_propostion}
\title{Find closest object names}
\usage{
error_correction_propostion(asked_objet, method = "jaccard", n = 2)
}
\arguments{
\item{asked_objet}{The R name producing an error.}

\item{method}{Method for distance calculation. The default is "jaccard",
see [stringdist::stringdist-metrics].}

\item{n}{Number of corrections to suggest (default: 2).}
}
\value{
A character vector with the closest neighbors, or `character(0)` if
  no valid input is provided.
}
\description{
Analyse a typo and suggests the closest names.
}
\examples{
error_correction_propostion("iri")
error_correction_propostion("view")
}
