% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_package.R
\docType{package}
\name{fcuk-package}
\alias{fcuk-package}
\alias{fcuk}
\title{fcuk}
\description{
Automatically suggests a correction when a typo occurs.
}
\author{
vincent <vincent@thinkr.fr>
}
