% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{get_all_objets_from_r}
\alias{get_all_objets_from_r}
\title{extract all object from r}
\usage{
get_all_objets_from_r(...)
}
\arguments{
\item{...}{not used}
}
\description{
fetch all r loaded objects
}
\examples{
get_all_objets_from_r()
}
