% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{get_all_objets_from_r}
\alias{get_all_objets_from_r}
\title{Extract the name of all objects loaded in the R environments}
\usage{
get_all_objets_from_r()
}
\value{
a list with the names of all objects contained in the environments
}
\description{
Fetch the name of all objects loaded in the environments (functions, values, data...)
}
\examples{
get_all_objets_from_r()
}
