% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun.R
\name{get_last}
\alias{get_last}
\title{Return corrected instruction}
\usage{
get_last(n = 1)
}
\arguments{
\item{n}{Position of the correction (default: 1).}
}
\value{
A character string with the corrected command, or `NA` if unavailable.
}
\description{
Get the corrected command at position `n`.
}
\examples{
get_last()
}
