% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{init_error_tracker}
\alias{init_error_tracker}
\title{Init error tracker}
\usage{
init_error_tracker()
}
\description{
After lauching this function, every error message will be analysed.
This function is called when loading the package.
}
\examples{
getOption("error")
fcuk::init_error_tracker()
getOption("error")
}
