% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun.R
\name{+.fcuk}
\alias{+.fcuk}
\title{Rerun with correction}
\usage{
\method{+}{fcuk}(x, y)
}
\arguments{
\item{x}{Blank object of class fcuk.}

\item{y}{An integer corresponding to the position of the proposal to be reused.}
}
\value{
Invisibly returns a blank object.
}
\description{
Use the `+` operator with the `.` object to apply a correction.
}
\examples{
\dontrun{
view(iris) # error
. + 1 # returns View(iris)
}
}
